refactor before adding tests
